+++ gradle/publishAllToMavenLocal.sh
@@ -18,8 +18,8 @@ export IS_EXCLUDE_DEMOS=true
 && ./gradlew :growingio-webservice:circler:publishMavenAgentPublicationToMavenLocal \
 && ./gradlew :growingio-tools:crash:publishMavenAgentPublicationToMavenLocal \
 && ./gradlew :growingio-tools:oaid:publishMavenAgentPublicationToMavenLocal \
-&& ./gradlew :growingio-adapter:analytics:publishMavenAgentPublicationToMavenLocal \
-&& ./gradlew :growingio-adapter:analytics3:publishMavenAgentPublicationToMavenLocal \
+&& ./gradlew :growingio-adapter:analytics-fa:publishMavenAgentPublicationToMavenLocal \
+&& ./gradlew :growingio-adapter:analytics-ga:publishMavenAgentPublicationToMavenLocal \
 && ./gradlew :gio-sdk:tracker:publishMavenAgentPublicationToMavenLocal \
 && ./gradlew :gio-sdk:tracker-cdp:publishMavenAgentPublicationToMavenLocal \
 && ./gradlew :gio-sdk:autotracker:publishMavenAgentPublicationToMavenLocal \
