+++ growingio-adapter/analytics-fa/.gitignore
