+++ growingio-adapter/analytics-fa/build.gradle
