+++ growingio-adapter/analytics-fa/gradle.properties
@@ -1,0 +1,4 @@
+POM_NAME=analytics-fa
+POM_ARTIFACT_ID=analytics-fa
+POM_PACKAGING=aar
+POM_DESCRIPTION=Adapter Google Analytics data to Growingio's sdk.
