+++ growingio-adapter/analytics-fa/proguard-rules.pro
