+++ growingio-adapter/analytics-fa/src/main/java/com/growingio/android/analytics/firebase/FirebaseAnalyticsLibraryModule.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (C) 2020 Beijing Yishu Technology Co., Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.growingio.android.analytics.firebase;
+
+import android.content.Context;
+
+import com.growingio.android.sdk.LibraryGioModule;
+import com.growingio.android.sdk.track.modelloader.TrackerRegistry;
+import com.growingio.sdk.annotation.GIOLibraryModule;
+
+/**
+ * <p>
+ *
+ * @author cpacm 2022/5/19
+ */
+@GIOLibraryModule
+public class FirebaseAnalyticsLibraryModule extends LibraryGioModule {
+    @Override
+    public void registerComponents(Context context, TrackerRegistry registry) {
+        //just init
+        FirebaseAnalyticsAdapter.init(context);
+        //registry.register(this.getClass(), Void.class, null);
+    }
+}
