+++ growingio-adapter/analytics-ga/.gitignore
