+++ growingio-adapter/analytics-ga/gradle.properties
@@ -1,0 +1,4 @@
+POM_NAME=analytics-ga
+POM_ARTIFACT_ID=analytics-ga
+POM_PACKAGING=aar
+POM_DESCRIPTION=Adapter Google Analytics data to Growingio's sdk.
