+++ growingio-adapter/analytics-ga/proguard-rules.pro
