+++ growingio-adapter/analytics-ga/src/androidTest/java/com/example/analytics3/ExampleInstrumentedTest.java
