--- growingio-adapter/analytics/gradle.properties
@@ -1,4 +1,0 @@
-POM_NAME=analytics
-POM_ARTIFACT_ID=analytics
-POM_PACKAGING=aar
-POM_DESCRIPTION=Adapter Google Analytics data to Growingio's sdk.
