--- growingio-adapter/analytics3/gradle.properties
@@ -1,4 +1,0 @@
-POM_NAME=analytics3
-POM_ARTIFACT_ID=analytics3
-POM_PACKAGING=aar
-POM_DESCRIPTION=Adapter Google Analytics data to Growingio's sdk.
