+++ settings.gradle
@@ -27,11 +27,12 @@ include ':growingio-tools:crash'
 include ':growingio-tools:snappy'
 include ':growingio-tools:oaid'
 
-include ':growingio-adapter:analytics'
+include ':growingio-adapter:analytics-fa'
+include ':growingio-adapter:analytics-ga'
 
 // sdks
 include ':gio-sdk:tracker'
 include ':gio-sdk:autotracker'
 include ':gio-sdk:tracker-cdp'
 include ':gio-sdk:autotracker-cdp'
-include ':growingio-adapter:analytics3'
+
